fix(api-server): isolate voting PDF-report tests into their own binary (BIT-158)
The `dev` Backend workflow has been red for hours because two tests in
`voting_tests.rs` (`pdf_report::test_get_report_pdf_*`) fail deterministically:
`POST /api/v1/auth/register` 500s with `relation "users" does not exist`,
so `create_authenticated_user` gets 500 instead of 201.

Root cause: `voting_tests.rs` was the only api-server test binary mixing
schema-less `#[sqlx::test]` auth-rejection tests with schema-full
`#[sqlx::test(migrator = "db::MIGRATOR")]` tests. Under sqlx 0.9 that mix
leaves the `db::MIGRATOR` tests running against an un-migrated database (no
`users` table), while structurally-uniform all-`db::MIGRATOR` binaries such as
`auth_enumeration_tests` register users fine.

Fix: move the two PDF-report E2E tests into a dedicated `voting_report_tests.rs`
binary where every test uses `db::MIGRATOR`, and leave `voting_tests.rs` as a
uniform no-migrator binary. No production code changes.
